requirements.txt
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.1
celery==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.5
Django==5.2
django-celery-beat==2.8.0
django-environ==0.12.0
django-tailwind==4.0.1
django-timezone-field==7.1
idna==3.10
kombu==5.5.3
pika==1.3.2
prompt_toolkit==3.0.51
psycopg2==2.9.10
python-crontab==3.2.0
python-dateutil==2.9.0.post0
requests==2.32.3
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.4.0
vine==5.1.0
wcwidth==0.2.13
